Use annotatedFunctionToChatCompletionFunction from shared
