+++ CopilotKit/packages/cloud/src/lib/openai-adapter.ts
@@ -1,7 +1,7 @@
 import OpenAI from "openai";
 import { CopilotKitServiceAdapter } from "../types/service-adapter";
 import { limitOpenAIMessagesToTokenCount, maxTokensForOpenAIModel } from "../utils/openai";
-import { AnnotatedFunction } from "@copilotkit/shared";
+import { AnnotatedFunction, annotatedFunctionToChatCompletionFunction } from "@copilotkit/shared";
 import { openaiStreamInterceptor } from "../utils";
 
 const DEFAULT_MODEL = "gpt-4-1106-preview";
@@ -37,37 +37,4 @@ export class OpenAIAdapter implements CopilotKitServiceAdapter {
       .toReadableStream();
     return openaiStreamInterceptor(stream, functions);
   }
-}
-
-// TODO proper type, maybe put it in shared
-function annotatedFunctionToChatCompletionFunction(
-  annotatedFunction: AnnotatedFunction<any[]>,
-): OpenAI.ChatCompletionCreateParams.Function {
-  // Create the parameters object based on the argumentAnnotations
-  let parameters: { [key: string]: any } = {};
-  for (let arg of annotatedFunction.argumentAnnotations) {
-    // isolate the args we should forward inline
-    let { name, required, ...forwardedArgs } = arg;
-    parameters[arg.name] = forwardedArgs;
-  }
-
-  let requiredParameterNames: string[] = [];
-  for (let arg of annotatedFunction.argumentAnnotations) {
-    if (arg.required) {
-      requiredParameterNames.push(arg.name);
-    }
-  }
-
-  // Create the ChatCompletionFunctions object
-  let chatCompletionFunction: OpenAI.ChatCompletionCreateParams.Function = {
-    name: annotatedFunction.name,
-    description: annotatedFunction.description,
-    parameters: {
-      type: "object",
-      properties: parameters,
-      required: requiredParameterNames,
-    },
-  };
-
-  return chatCompletionFunction;
 }
