+++ CopilotKit/packages/react-core/src/components/copilot-provider/copilot-provider.tsx
@@ -5,7 +5,11 @@ import { CopilotContext, CopilotApiConfig } from "../../context/copilot-context"
 import useTree from "../../hooks/use-tree";
 import { ChatCompletionCreateParams } from "openai/resources/chat";
 import { DocumentPointer } from "../../types";
-import { FunctionCallHandler, AnnotatedFunction, Function } from "@copilotkit/shared";
+import {
+  FunctionCallHandler,
+  AnnotatedFunction,
+  annotatedFunctionToChatCompletionFunction,
+} from "@copilotkit/shared";
 import useFlatCategoryStore from "../../hooks/use-flat-category-store";
 import { StandardCopilotApiConfig } from "./standard-copilot-api-config";
 import { CopilotProviderProps } from "./copilot-provider-props";
@@ -227,36 +231,4 @@ function entryPointsToChatCompletionFunctions(
   entryPoints: AnnotatedFunction<any[]>[],
 ): ChatCompletionCreateParams.Function[] {
   return entryPoints.map(annotatedFunctionToChatCompletionFunction);
-}
-
-function annotatedFunctionToChatCompletionFunction(
-  annotatedFunction: AnnotatedFunction<any[]>,
-): Function {
-  // Create the parameters object based on the argumentAnnotations
-  let parameters: { [key: string]: any } = {};
-  for (let arg of annotatedFunction.argumentAnnotations) {
-    // isolate the args we should forward inline
-    let { name, required, ...forwardedArgs } = arg;
-    parameters[arg.name] = forwardedArgs;
-  }
-
-  let requiredParameterNames: string[] = [];
-  for (let arg of annotatedFunction.argumentAnnotations) {
-    if (arg.required) {
-      requiredParameterNames.push(arg.name);
-    }
-  }
-
-  // Create the ChatCompletionFunctions object
-  let chatCompletionFunction: Function = {
-    name: annotatedFunction.name,
-    description: annotatedFunction.description,
-    parameters: {
-      type: "object",
-      properties: parameters,
-      required: requiredParameterNames,
-    },
-  };
-
-  return chatCompletionFunction;
 }
