+++ CopilotKit/packages/shared/src/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./utils";
 export * from "./parse-chat-completion";
 export * from "./decode-chat-completion";
 export * from "./decode-chat-completion-as-text";
+export * from "./annotated-function";
